Remove `config_vars` from bin/release
This field is not currently documented at https://devcenter.heroku.com/articles/buildpack-api#bin-release, but it is used for setting config vars for customers. The caveat is that it only affects apps on the first push, which isn't very useful. That's why it's not even documented anymore.

The "correct" way to set an environment variable is:

- `.profile.d/<script>.sh` Sets the value at launch/runtime
- `export` Sets the value at build time for future buildpacks

Previously, the `def default_config_vars` was used in relationship with `bin/release` to set `config_vars`. But, since that's not used (as of this commit) it is repurposed as an interface for writing default values to `.profile.d` scripts. All these values were already being written, but now it's centralized (i.e. this .profile.d part of the change is a literal refactor, it doesn't change any behavior).

There are some implications with the API: Any value that comes out of `def default_config_vars` will be persisted to disk and will take precedent if someone runs `heroku config:unset <env var>` for example. Previously, the values also contained the user's environment variable value i.e. if someone set `heroku config:set RAILS_ENV=my_custom_env` then that change would be propagated to `bin/release`. Now we don't want to write a value like  `        export RACK_ENV=${RACK_ENV:-my_custom_env}` otherwise. We want the default to fall back to OUR default. Therefore only literal values should be used in this API. That part is unexpected and unsafe, so I'm likely going to move/change how the interface works. But I will do that in future commits.