Remove set empty env var
This env var was set for Bundler <= 2.1.4 compat in this commit https://github.com/heroku/heroku-buildpack-ruby/commit/655e246657e329db524177d410e4ae4aaf8cc141 which explains the behavior.

It has since been reduced over the years to no longer set it to a `1` and instead set it to an empty string. It's also only exported and not set for build time, therefore it can be safely removed.

As to **why** it isn't a problem any more. We list our oldest available Ruby version as 3.1:

https://devcenter.heroku.com/articles/ruby-support-reference#unsupported-ruby-versions

And the default bundler version for that version of ruby is 2.3 which is > 2.1

```
$ gem list | grep bundle
bundler (2.6.9, 2.5.22, default: 2.3.27)
```