Remove test for behavior no longer present
The `bundle platform --ruby` codepath will return a version even if it's not in the Gemfile.lock, this "feature" was heavilly used by tests for version switching.

When the GemfileLock method of Ruby version detection was put into production, we also used metrics to see if this codepath is called in the wild, and it's exceedingly rare.