Support Bundler 4
Bundler 4 is the most recent version and will be coupled with Ruby 4, which will be released on December 25th.

The logic for "blessed" bundler versions is in flux as we work towards matching the CNB behavior of installing the customer's exact specified version in their Gemfile.

Some of that work is staged in https://github.com/heroku/heroku-buildpack-ruby/pull/1665.