Remove-unneeded call (#1666)
Previously the `new_app?` logic was lazy and memoized, we called it at the start of compile to make sure it checked the cache dir before anything manipulated it. That logic has been moved out to `def initialize` and sets a `@new_app` ivar such that this call here does nothing.