Ruby version detection now uses Gemfile.lock contents instead of `bundle platform --ruby` (#1665)
* Explicit order of ruby_version evaluation

The prior logic of `ruby_version` would lazy initialize making the order and location of evaluation difficult to follow. This change makes the logic and location explicit.

* Remove reliance on bundle_platform_ruby in tests

The `RubyVersion.bundle_platform_ruby` method of deriving the RubyVersion requires shelling out to `bundle platform --ruby` which requires bundler is installed BEFORE we determine what ruby version to use. The new GemfileLock parsing logic does not have this same requirement. Moving tests to rely on that logic allows us to delay installing bundler (eventually) until after a Ruby version is installed.

* Remove reliance on bundle_platform_ruby

The `RubyVersion.bundle_platform_ruby` method of deriving the RubyVersion requires shelling out to `bundle platform --ruby` which requires bundler is installed BEFORE we determine what ruby version to use. The new GemfileLock parsing logic does not have this same requirement. Moving tests to rely on that logic allows us to delay installing bundler (eventually) until after a Ruby version is installed.

* Remove reliance on bundle_platform_ruby

The `RubyVersion.bundle_platform_ruby` method of deriving the RubyVersion requires shelling out to `bundle platform --ruby` which requires bundler is installed BEFORE we determine what ruby version to use. The new GemfileLock parsing logic does not have this same requirement. Moving tests to rely on that logic allows us to delay installing bundler (eventually) until after a Ruby version is installed.

* Remove unused Bundler.with_unbundled_env

This was needed due to invoking bundler from within a project using bundler. As we're no longer invoking bundler, it's no longer needed

* Remove BundlerWrapper.ruby_version

This method requires shelling out to `bundle platform --ruby` which requires bundler is installed BEFORE we determine what ruby version to use. The new GemfileLock parsing logic does not have this same requirement.

* Update test to use Gemfile.lock to specify Ruby version

* Update test to use Gemfile.lock to specify Ruby version

* Remove test for behavior no longer present

The `bundle platform --ruby` codepath will return a version even if it's not in the Gemfile.lock, this "feature" was heavilly used by tests for version switching.

When the GemfileLock method of Ruby version detection was put into production, we also used metrics to see if this codepath is called in the wild, and it's exceedingly rare.

* Changelog