Remove legacy jruby PATH
Added in https://github.com/heroku/heroku-buildpack-ruby/commit/74b428420bfbc5f69cf489836e53d16ce1abdb74 this was when the stack image included a copy of java, which it no longer does. The local `bin` dir is put on the path in the general case but this was never refactored away.