Install Ruby before bundler (#1676)
* Move ruby version logic to class method

The goal is to install Ruby and Bundler versions before instantiating any "language pack" because the `detect` logic checks if specific gems are installed. By moving moving this logic to a class method, we can decouple it from the detection lifecycle.

* Move ruby version logic to class method

The goal is to install Ruby and Bundler versions before instantiating any "language pack" because the `detect` logic checks if specific gems are installed. By moving moving this logic to a class method, we can decouple it from the detection lifecycle.

* Move ruby version logic to class method

The goal is to install Ruby and Bundler versions before instantiating any "language pack" because the `detect` logic checks if specific gems are installed. By moving moving this logic to a class method, we can decouple it from the detection lifecycle.

* Inline logic

* Inline bundler binstubs path

* Inline variable

* Reduce injected values

Both `ruby_layer_path` and `gem_layer_path` point at the same value which is better defined as "app_path".

* ENV = Pathname#to_s

* Move yarn logic to class

* Inject ruby version

* Inject ruby install path

* Inject default config vars

* Inject user_env_hash

* Make class method

* Move logic to class method

* Move database yaml logic

The syntax highlighting  doesn't like there combination of heredoc and ERB, moving it is an ergonomic QOL improvement for developers using an IDE with syntax highlighting.

* Inject values into method

* Inline shelling out logic

* Remove leftover ivar

* Make class method

* Inject app path and io

* Inject bundler_cache

* Inject bundler version

* Prefer app_path over Dir.pwd

* Inject IO

* Remove unused variable

* Class method

* Don't need to track differences anymore

* Add Metadata#try_read

* Simplify metadata logic

* Reduce variable usage

* Remove unused code

* Inline variable

* Refactor

* Fixup missing key

* Remove unused logic

* Inject IO

* Call pipe on self instead of IO

* Call on self instead of IO

* Extract warning container

* Remove confusing method name

* Support installing with an absolute path

* Move logic

* Align setup_export test with main Ruby

* Refactor ENV logic into class

* Remove legacy jruby PATH

Added in https://github.com/heroku/heroku-buildpack-ruby/commit/74b428420bfbc5f69cf489836e53d16ce1abdb74 this was when the stack image included a copy of java, which it no longer does. The local `bin` dir is put on the path in the general case but this was never refactored away.

* Remove unused kwarg

* Remove allow_git

This is no longer needed

* Extract logic to single class method

* Extract bundle_list

* Refactor logic to not need shelling out

* Replace slug_vendor_base with ruby_version logic

* Replace slug_vendor_base

* Replace slug_vendor_base

* Replace slug_vendor_base

* Replace slug_vendor_base

* Dead code

* Remove slug_vendor_ base

* "build_dir" -> "app_path" for consistency

* Remove bundler instantiation from `test` command

* Dead code

* Inline method

* Move outdated version check

* Dependency inject bundler

* Refactor

* Fix test compile

* Inject Ruby Version

* Extract WarnIO

* Fix

* Inject new_app

* Extract arch

* Fix new_app

* Move code

* Inject "bundler"

* Expose default config vars logic as class method

* Refactor logic out of instance

* Remove instance method, delegate to class method

* Inject bundler

* Inline logic

* Remove class method, inline logic, reorder

* Move variable assignment

* Standardize class method calling convention

* Changelog