Don't export bootstrap Ruby on PATH (And Fix "herokuish" CI test case)
Previously the Ruby buildpack worked like this:

- Install a "bootstrap" version of Ruby
- Use bootstrap ruby to execute the buildpack code (written in ruby)
- Download customer's bundler version
- Run that bundler version with bootstrap ruby to determine customer's ruby version `bundle platform --ruby`
- ...

With that flow, it's required that the Ruby program must be able to re-exec itself since it must have a Ruby version to run and a ruby version to run `bundle platform --ruby`.

The code was changed such that it now looks like this:

- Install a "bootstrap" version of Ruby
- Use bootstrap ruby to execute the buildpack code (written in ruby)
- Statically determine Ruby version from the Gemfile.lock
- Install customer's Ruby version
- Install customer's bundler version
- ...

With this flow, we do not ever need to re-exec the bootstrap ruby version. This lets us remove exporting it to the PATH which has caused a LOT of problems over the years.

As mentioned it also fixes the "herokuish" bug that was introduced in v334 (then rolled back). I was able to reproduce the bug via a test added in https://github.com/heroku/heroku-buildpack-ruby/pull/1687. The test uses a different version of Ruby than the "bootstrap" version. The problem was this error:

```
Running: which -a ruby
  /app/bin/ruby
  /tmp/tmp.Dx6ifBJLoP/bin//ruby
  /app/bin/ruby
Running: which -a bundle
  /tmp/tmp.Dx6ifBJLoP/bin//bundle
  /app/vendor/bundle/bin/bundle
  /app/vendor/bundle/ruby/3.1.0/bin/bundle
Running: bundle list
/tmp/buildpacks/5332b9a13007e20408c130cf3c47b025a8417ef0/lib/language_pack/helpers/bundle_list.rb:56:in `call': Error detecting dependencies (LanguagePack::Helpers::BundleList::CmdError)

The Ruby buildpack requires information about your application???s dependencies to
complete the build. Without this information, the Ruby buildpack cannot continue.

Command failed: `bundle list`

Could not find sinatra-4.2.1, puma-7.1.0, rack-3.2.4, rake-13.3.1, rackup-2.3.1, rack-test-2.2.0, test-unit-3.7.3, logger-1.7.0, mustermann-3.0.4, rack-protection-4.2.1, rack-session-2.1.1, tilt-2.6.1, nio4r-2.7.4, power_assert-3.0.1, base64-0.3.0 in locally installed gems
Install missing gems with `bundle install`.
```

(Debugging output added for later explanation)

This comes from a Heroku CI run which works by executing `bin/test-compile` (installing ruby and bundler etc.) and then calling `bin/test`. The `bin/test` already has a full ruby + bundler environment setup and ready to go, but we still download a bootstrap version of ruby so that the code only ever has to support one Ruby version (rather than N versions across the lifecycle of M stacks).

What's happening in the output is that we see that `ruby` being used is `/app/bin/ruby` (local `./bin` binstub) because we forced it onto the path via: 

```
LanguagePack::ShellHelpers.user_env_hash["PATH"] = "#{app_path.join("bin")}:#{ENV["PATH"]}"
```

However the `bundle` executable is pointing to `/tmp/tmp.Dx6ifBJLoP/bin//bundle` instead of the correct `/app/vendor/bundle/bin/bundle` because we don't symlink that to the binstub directory (people expect to check in their `./bin/bundle` binstub and not have it overwritten). And because the PATH has the bootstrap ruby directory `/tmp/tmp.Dx6ifBJLoP/bin/` and that also includes a `bundle` executable, it uses that to call `gem list`.

Why weren't all of the tests failing? Well, the gems are installed to `./vendor/bundle/ruby/<major>.<minor>.0/`. The bundler that ships with ruby doesn't use PATH but instead is hardcoded to use the version of ruby it was shipped with:

```
$ chruby 3.3.9
$ cat $HOME/.rubies/ruby-3.3.9/bin/bundle
#!/Users/rschneeman/.rubies/ruby-3.3.9/bin/ruby
```

Or

```
$ curl https://heroku-buildpack-ruby.s3.us-east-1.amazonaws.com/heroku-24/arm64/ruby-4.0.0.preview3-bb0a79e.tgz -O
$ ...
$ cat ruby-4.0.0.preview3-bb0a79e/bin/bundle
#!/bin/sh
# -*- ruby -*-
# This file was generated by RubyGems.
#
# The application 'bundler' is installed as part of a gem, and
# this file is here to facilitate running it.
#
_=_\
=begin
bindir="${0%/*}"
exec "$bindir/ruby" "-x" "$0" "$@"
=end
#!/usr/bin/env ruby
```

Now, bundler is smart enough to look in `./vendor/bundle/ruby/<major>.<minor>.0/` when BUNDLE_DEPLOYMENT is set, BUT since it's being executed with the bootstrap ruby version, it will always be looking in `./vendor/bundle/ruby/3.3.0/` (bootstrap ruby version is 3.3.9 now). The test fails because it's using Ruby 3.1.3 which has gems installed to `./vendor/bundle/ruby/3.1.0/` which doesn't exist, so `gem list` fails to find any of the gems (because it's looking at the wrong directory because it was accidentally executed with the wrong ruby version.

Previously the fix was to exploit the "user provided environment variables" for a purpose it was never intended for, to put the customer's PATH first (in front of the bootstrap ruby version):

```
# - Add bundler's bin directory to the PATH
# - Always make sure `$HOME/bin` is first on the path
user_env_hash["PATH"] = "#{build_dir}/bin:#{bundler.bundler_path}/bin:#{user_env_hash["PATH"]}"
user_env_hash["GEM_PATH"] = LanguagePack::Ruby.slug_vendor_base
```

This approach requires knowing the bundler_path which is now derived like this:

```
  # ...
  class RubyVersion
    # Ruby versioned bundler directory
    #
    # When installing gems via `BUNDLE_DEPLOYMENT=1 bundle install`, they're installed into a versioned directory based on the ruby version.
    #
    # This becomes the location of GEM_PATH on disk https://www.schneems.com/2014/04/15/gem-path.html.
    # - Executables are at bundler_directory.join("bin")
    # - Gems are at bundler_directory.join("gems")
    #
    # For example:
    #
    # - Ruby 3.4.7 would be "vendor/bundle/ruby/3.4.0"
    # - JRuby 9.4.14.0 would be "vendor/bundle/jruby/3.1.0" (As it implements Ruby 3.1.7 spec)
    def bundler_directory
      "vendor/bundle/#{engine}/#{major}.#{minor}.0"
    end
```

OR previously it was the same as `slug_vendor_base` which is derived by shelling out:

```
  # For example "vendor/bundle/ruby/2.6.0"
  def self.slug_vendor_base
    @slug_vendor_base ||= begin
      command = %q(ruby -e "require 'rbconfig';puts \"vendor/bundle/#{RUBY_ENGINE}/#{RbConfig::CONFIG['ruby_version']}\"")
      out = run_no_pipe(command, user_env: true).strip
      error "Problem detecting bundler vendor directory: #{out}" unless $?.success?
      out
    end
  end
```

A better approach such as un-shifting the first value of path, or passing the bootstrap ruby version wasn't really feasible due to the architecture of the program. The refactoring that happened in v334 (that is reverted and re-applied in this PR) enabled this seemingly small, but very impactful simplification (removing re-exporting of the bootstrap PATH).