Modify CI test to exercise `bin/test` code
Previously this test worked by using the `app.json` to define a test script. When that happens, the `bin/test` code is completely bypassed, so it's not under test.

This new modification instead uses the `bin/test` logic to call `bin/rake test` and that file prints the same information using the same script.

While the assertions in the test on this branch didn't need to change (except for asserting `bin/rake` is now in the `bin` directory), they will fail on `main` due to the way that `bin/test` is implemented, it resolves path order differently which can cause different execution behavior than if you define a test via `app.json` and different than the normally deployed Heroku app (in production).